config patch & ace paradrop fix
