+++ addons/orbis_ace3_cargo_config/CfgEventHandlers.hpp
@@ -1,0 +1,11 @@
+class Extended_PreStart_EventHandlers {
+	class orbis_ace3_cargo_config {
+		init = "call compile preProcessFileLineNumbers 'orbis_ace3_cargo_config\XEH_preStart.sqf'";
+	};
+}; 
+
+class Extended_PreInit_EventHandlers {
+	class orbis_ace3_cargo_config {
+		init = "call compile preProcessFileLineNumbers 'orbis_ace3_cargo_config\XEH_preInit.sqf'";
+	};
+};
