+++ addons/orbis_ace3_cargo_config/XEH_PREP.sqf
@@ -1,0 +1,5 @@
+// Original
+
+
+// ACE3 orverload
+ace_cargo_fnc_paradropItem = compile preprocessFileLineNumbers "orbis_ace3_cargo_config\scripts\fnc_paradropItem_config.sqf";
