+++ addons/orbis_ace3_cargo_config/config.cpp
@@ -1,0 +1,12 @@
+class CfgPatches {
+	class orbis_ace3_cargo_config {
+		name = "Orbis ACE3 Cargo Config";
+		author = "Orbis2358";
+		requiredVersion = 0.1;
+		requiredAddons[] = {"ace_main", "ace_cargo"};
+		units[] = {};
+		weapons[] = {};
+	};
+};
+
+#include "CfgEventHandlers.hpp"
