--- addons/orbis_ace3_medical_config/CfgPatches.hpp
@@ -1,10 +1,0 @@
-class CfgPatches {
-	class orbis_ace3_medical_config {
-		name = "Orbis ACE3 Medical Config";
-		author = "Orbis2358";
-		requiredVersion = 0.1;
-		requiredAddons[] = {"ace_main", "ace_medical", "ace_medical_menu"};
-		units[] = {};
-		weapons[] = {};
-	};
-};
