+++ addons/orbis_ace3_medical_config/config.cpp
@@ -1,5 +1,13 @@
-#include "CfgPatches.hpp"
+class CfgPatches {
+	class orbis_ace3_medical_config {
+		name = "Orbis ACE3 Medical Config";
+		author = "Orbis2358";
+		requiredVersion = 0.1;
+		requiredAddons[] = {"ace_main", "ace_medical", "ace_medical_menu"};
+		units[] = {};
+		weapons[] = {};
+	};
+};
 
 #include "ACE3_Medical_Treatments.hpp"
-
 #include "CfgEventHandlers.hpp"
